refactor: test file ttoklip-stresstest/sources로 이동
